+++ arch/src/riscv64/kcontext.rs
@@ -1,0 +1,74 @@
+use riscv::register::{
+    satp::{self, Satp},
+    sepc,
+    sstatus::{self, Sstatus},
+};
+
+use crate::memory::TLB;
+
+/// Store some permission flags
+#[derive(Debug)]
+pub struct KContext {
+    // For preempt and non preempt
+    /// Permit supervisor user memory access
+    sum_cnt: usize,
+
+    // For preempt only
+    sstatus: Sstatus,
+    sepc: usize,
+    satp: Satp,
+}
+
+impl KContext {
+    pub const fn new() -> Self {
+        Self {
+            sum_cnt: 0,
+            sstatus: Sstatus::from_bits(0),
+            sepc: 0,
+            satp: Satp::from_bits(0),
+        }
+    }
+
+    pub unsafe fn auto_sum(&self) {
+        log::trace!("[KContext::auto_sum] sum_cnt: {}", self.sum_cnt);
+        if self.sum_cnt == 0 {
+            unsafe { riscv::register::sstatus::clear_sum() };
+        } else {
+            unsafe { riscv::register::sstatus::set_sum() };
+        }
+    }
+
+    pub fn inc_sum(&mut self) {
+        if self.sum_cnt == 0 {
+            unsafe { riscv::register::sstatus::set_sum() };
+        }
+        self.sum_cnt += 1;
+    }
+
+    pub fn dec_sum(&mut self) {
+        debug_assert!(self.sum_cnt > 0);
+        self.sum_cnt -= 1;
+        if self.sum_cnt == 0 {
+            unsafe { riscv::register::sstatus::clear_sum() };
+        }
+    }
+
+    pub fn change_env(&self, new: &Self) {
+        unsafe { new.auto_sum() };
+    }
+
+    pub fn preempt_record(&mut self) {
+        self.sstatus = sstatus::read();
+        self.sepc = sepc::read();
+        self.satp = satp::read();
+    }
+
+    pub unsafe fn preempt_resume(&self) {
+        unsafe {
+            sstatus::write(self.sstatus);
+            sepc::write(self.sepc);
+            satp::write(self.satp);
+            TLB::sfence_vma_all();
+        }
+    }
+}
