+++ arch/src/riscv64/memory/mod.rs
@@ -9,14 +9,18 @@ pub use frame::*;
 pub use page_table::PageTable;
 pub use pte::PageTableEntry;
 
-pub unsafe fn sfence_vma_vaddr(vaddr: usize) {
-    unsafe {
-        core::arch::riscv64::sfence_vma_vaddr(vaddr);
-    }
-}
+pub struct TLB;
 
-pub unsafe fn sfence_vma_all() {
-    unsafe {
-        core::arch::riscv64::sfence_vma_all();
+impl TLB {
+    pub unsafe fn sfence_vma_vaddr(vaddr: usize) {
+        unsafe {
+            core::arch::riscv64::sfence_vma_vaddr(vaddr);
+        }
+    }
+
+    pub unsafe fn sfence_vma_all() {
+        unsafe {
+            core::arch::riscv64::sfence_vma_all();
+        }
     }
 }
